Final touches before initiating Pull Request
- renamed 'landslide.py' as 'landslide_probability.py'.
- removed drivers for testing from landslides folder.
